requirements.txt
Django==1.7.1
djangorestframework==3.0.2
requests==0.13.2
stripe==1.19.1
wsgiref==0.1.2
